+++ example/lib/custom_model_directory/article/custom_base_directory_obj.g.dart
@@ -14,7 +14,15 @@ CustomBaseDirectoryObj _$CustomBaseDirectoryObjFromJson(
 }
 
 Map<String, dynamic> _$CustomBaseDirectoryObjToJson(
-        CustomBaseDirectoryObj instance) =>
-    <String, dynamic>{
-      'name': instance.name,
-    };
+    CustomBaseDirectoryObj instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull('name', instance.name);
+  return val;
+}