+++ example/lib/model/user/person/person.dart
@@ -7,9 +7,13 @@ part 'person.g.dart';
 @JsonSerializable(explicitToJson: true)
 @immutable
 class Person {
-  @JsonKey(name: 'firstName', required: true)
+  @JsonKey(name: 'firstName', required: true, includeIfNull: false)
   final String firstName;
-  @JsonKey(name: 'gender', required: true, unknownEnumValue: Gender.X)
+  @JsonKey(
+      name: 'gender',
+      required: true,
+      includeIfNull: false,
+      unknownEnumValue: Gender.X)
   final Gender gender;
 
   const Person({